♻️ Make Pf2eSpeed.getSpeed into a Pf2eJsonNodeReader default method
